requirements.txt
# Core ML & DL
torch>=2.3,<2.5
accelerate>=0.27,<0.29
bitsandbytes>=0.39.0
# Include Run:ai loader extras for S3 streaming; pin to a stable 0.10.x release
vllm[runai]>=0.10.1,<0.11.0
runai-model-streamer>=0.1.0

# Transformers & NLP
transformers>=4.42,<4.46
sentence-transformers>=2.7,<2.8

# Data & Analysis
numpy>=1.24.0
pandas>=1.5.0
matplotlib>=3.6.0
seaborn>=0.12.0

# Utilities
wikipedia-api>=0.5.4
requests>=2.28.0
tqdm>=4.65.0
pytest>=7.4,<8.1

# AWS / Postgres
boto3>=1.35.0
sqlalchemy>=2.0.0,<2.1.0
psycopg2-binary>=2.9.0
pgvector>=0.2.4

# Infra / CDK
aws-cdk-lib==2.173.0
constructs>=10.3.0,<11.0.0
PyYAML>=6.0
